Birthday module

+++ app/Http/Controllers/CMSController.php
@@ -8,6 +8,7 @@ use App\Models\Coupon;
 use App\Models\Course;
 use App\Models\Lesson;
 use App\Models\Section;
+use App\Models\Video;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Crypt;
@@ -45,51 +46,50 @@ class CMSController extends Controller
 
     public function save_course($id, Request $request)
     {
-        // dd($request->course_description);
-
-        // switch($request->submitbutton) {
-
-        //     case 'save-draft': 
-        //         //action save here
-        //         return back()->with([
-        //             'type' => 'success',
-        //             'message' => 'Saved'
-        //         ]);
-        //     break;
-        
-        //     case 'save-review': 
-        //         //action for save-draft here
-        //         return back()->with([
-        //             'type' => 'success',
-        //             'message' => 'Confirm'
-        //         ]);
-        //     break;
-        // }
-
         $finder = Crypt::decrypt($id);
 
         $course = Course::find($finder);
 
         $category = Category::find($request->category);
 
-        if($request->price == null)
+        if($request->courseUpdate)
         {
-            $price = 0;
-        } else {
-            $price = $request->price;
-        }
-
-        if (request()->hasFile('image')) 
-        {
-            $this->validate($request, [
-                'image' => 'required|mimes:jpeg,png,jpg',
-            ]);
-            $filename = request()->image->getClientOriginalName();
-            if($course->image) {
-                Storage::delete(str_replace("storage", "public", $course->image));
+            if($request->price == null)
+            {
+                $price = 0;
+            } else {
+                $price = $request->price;
             }
-            request()->image->storeAs('course_photo', $filename, 'public');
-            
+
+            if (request()->hasFile('image')) 
+            {
+                $this->validate($request, [
+                    'image' => 'required|mimes:jpeg,png,jpg',
+                ]);
+                $filename = request()->image->getClientOriginalName();
+                if($course->image) {
+                    Storage::delete(str_replace("storage", "public", $course->image));
+                }
+                request()->image->storeAs('course_photo', $filename, 'public');
+                
+                $course->update([
+                    'category_id' => $category->id,
+                    'title' => $request->title,
+                    'subtitle' => $request->subtitle,
+                    'description' => $request->course_description,
+                    'language' => $request->language,
+                    'level' => $request->level,
+                    'currency' => $request->currency,
+                    'price' => $price,
+                    'image' => '/storage/course_photo/'.$filename,
+                ]);
+
+                return back()->with([
+                    'type' => 'success',
+                    'message' => 'Save successfully.'
+                ]);
+            }
+
             $course->update([
                 'category_id' => $category->id,
                 'title' => $request->title,
@@ -99,8 +99,7 @@ class CMSController extends Controller
                 'image' => $request->image,
                 'level' => $request->level,
                 'currency' => $request->currency,
-                'price' => $price,
-                'image' => '/storage/course_photo/'.$filename,
+                'price' => $price
             ]);
 
             return back()->with([
@@ -109,23 +108,26 @@ class CMSController extends Controller
             ]);
         }
 
+         //Validate Request
+         $this->validate($request, [
+            'title' => ['required'],
+            'subtitle' => ['required'],
+            'course_description' => ['required'],
+            'language' => ['required'],
+            'level' => ['required'],
+            'currency' => ['required'],
+            'price' => ['required'],
+        ]);
 
         $course->update([
-            'category_id' => $category->id,
-            'title' => $request->title,
-            'subtitle' => $request->subtitle,
-            'description' => $request->course_description,
-            'language' => $request->language,
-            'image' => $request->image,
-            'level' => $request->level,
-            'currency' => $request->currency,
-            'price' => $price
+            'published' => true
         ]);
 
         return back()->with([
             'type' => 'success',
-            'message' => 'Save successfully.'
+            'message' => 'Published successfully.'
         ]);
+        
     }
 
     public function save_curriculum($id, Request $request)
@@ -151,59 +153,72 @@ class CMSController extends Controller
         ]);
     }
 
-    public function update_curriculum($id, Request $request)
+    public function action_curriculum($id, Request $request)
     {
-        //Validate Request
-        $this->validate($request, [
-            'code' => ['required', 'string'],
-            'percent' => ['required', 'numeric'],
-            'quantity' => ['required', 'numeric'],
-            'expires' => ['required', 'date'],
-        ]);
+        switch($request->submitbutton) {
 
-        $finder = Crypt::decrypt($id);
+            case 'Delete': 
+                //action save here
+                $finder = Crypt::decrypt($id);
 
-        $coupon = Coupon::find($finder);
+                $section = Section::find($finder);
 
-        $coupon->update([
-            'code' => $request->code,
-            'percent' => $request->percent,
-            'quantity' => $request->quantity,
-            'expires' => $request->expires
-        ]);
+                $lessons = Lesson::where('section_id', $section->id)->get();
 
-        return back()->with([
-            'type' => 'success',
-            'message' => 'Updated successfully.'
-        ]);
-    }
+                if($lessons)
+                {
+                    foreach($lessons as $lesson)
+                    {
+                        $video = Video::where('lesson_id', $lesson->id)->first();
+                        $token = explode('/', $video->original_filename);
+                        $token2 = explode('.', $token[sizeof($token)-1]);
+                
+                        if($video->original_filename)
+                        {
+                            cloudinary()->destroy(config('app.name').'/'.$token2[0]);
+                        }
 
-    public function delete_curriculum($id, Request $request)
-    {
-        //Validate Request
-        $this->validate($request, [
-            'delete_field' => ['required', 'string', 'max:255']
-        ]);
+                        $video->delete();
+                
+                        $lesson->delete();
+                    }
 
-        if($request->delete_field == "DELETE")
-        {
-            $finder = Crypt::decrypt($id);
+                }
 
-            Coupon::find($finder)->delete();
+                $section->delete();
 
-            return back()->with([
-                'type' => 'success',
-                'message' => 'Coupon deleted successfully.'
-            ]);
+                return back()->with([
+                    'type' => 'success',
+                    'message' => 'Section deleted successfully.'
+                ]);
+            break;
+        
+            case 'Update': 
+                //action for save-draft here
+
+                //Validate Request
+                $this->validate($request, [
+                    'section_title' => ['required', 'string', 'max:255']
+                ]);
+
+                $finder = Crypt::decrypt($id);
+
+                $section = Section::find($finder);
+
+                $section->update([
+                    'title' => $request->section_title,
+                    'objective' => $request->section_objective
+                ]);
+
+                return back()->with([
+                    'type' => 'success',
+                    'message' => 'Section updated successfully.'
+                ]);
+            break;
         }
-
-        return back()->with([
-            'type' => 'danger',
-            'message' => "Field doesn't match, Try Again!"
-        ]); 
     }
 
-    public function save_lesson($id, Request $request)
+    public function save_lesson(Request $request)
     {
         //Validate Request
         $this->validate($request, [
@@ -212,50 +227,74 @@ class CMSController extends Controller
             'content_type' => ['required', 'string', 'max:255'],
         ]);
 
-        $finder = Crypt::decrypt($id);
+        $section = Section::find($request->section_id);
 
-        $section = Section::find($finder);
-
-        if($$request->content_type == 'Video')
-        {
-            // $this->validate($request, [
-            //     'lesson_video' => ['required', 'string', 'max:255'],
-            // ]);
-
-            // Lesson::create([
-            //     'section_id' => $section->id,
-            //     'course_id' => $request->course_id,
-            //     'title' => $request->lesson_title,
-            //     'description' => $request->course_id,
-            //     'content_type' => $request->content_type,
-            //     'duration' => $request->lesson_duration,
-            // ]);
-
-        } elseif($$request->content_type == 'Youtube') 
+        if($request->content_type == 'Video')
         {
             $this->validate($request, [
-                'lesson_video' => ['required', 'string', 'max:255'],
+                'lesson_video' => 'required|mimes:mp4,mov,ogg,qt,wmv,avi,m3u8|max:20000',
             ]);
 
-            Lesson::create([
+            $file = request()->lesson_video->getClientOriginalName();
+                
+            $filename = pathinfo($file, PATHINFO_FILENAME);
+
+            $response = cloudinary()->uploadFile($request->file('lesson_video')->getRealPath(),
+                            [
+                                'folder' => config('app.name'),
+                                "public_id" => $filename,
+                                "use_filename" => TRUE
+                            ])->getSecurePath();
+            
+            $lesson = Lesson::create([
                 'section_id' => $section->id,
-                'course_id' => $request->course_id,
+                'course_id' => $section->course_id,
                 'title' => $request->lesson_title,
                 'description' => $request->course_id,
                 'content_type' => $request->content_type,
                 'duration' => $request->lesson_duration,
             ]);
 
+            Video::create([
+                'lesson_id' => $lesson->id,
+                'original_filename' => $response
+            ]);
+
             return back()->with([
                 'type' => 'success',
-                'message' => 'Section save successfully.'
+                'message' => 'Lesson save successfully.'
+            ]);
+
+        } elseif($request->content_type == 'Youtube') 
+        {
+            $this->validate($request, [
+                'lesson_youtube' => ['required', 'string', 'max:255'],
+            ]);
+
+            $lesson = Lesson::create([
+                'section_id' => $section->id,
+                'course_id' => $section->course_id,
+                'title' => $request->lesson_title,
+                'description' => $request->course_id,
+                'content_type' => $request->content_type,
+                'duration' => $request->lesson_duration,
+            ]);
+
+            Video::create([
+                'lesson_id' => $lesson->id,
+                'youtube_link' => $lesson->lesson_youtube
+            ]);
+
+            return back()->with([
+                'type' => 'success',
+                'message' => 'Lesson save successfully.'
+            ]);
+        } else {
+            return back()->with([
+                'type' => 'success',
+                'message' => "Content type doesn't exist on our database."
             ]);
         }
-
-        return back()->with([
-            'type' => 'success',
-            'message' => 'Section save successfully.'
-        ]);
     }
 
     public function update_lesson($id, Request $request)
@@ -285,29 +324,29 @@ class CMSController extends Controller
         ]);
     }
 
-    public function delete_lesson($id, Request $request)
+    public function delete_lesson($id)
     {
-        //Validate Request
-        $this->validate($request, [
-            'delete_field' => ['required', 'string', 'max:255']
-        ]);
+        $finder = Crypt::decrypt($id);
+        
+        $lesson = Lesson::findorfail($finder);
 
-        if($request->delete_field == "DELETE")
+        $video = Video::where('lesson_id', $lesson->id)->first();
+        $token = explode('/', $video->original_filename);
+        $token2 = explode('.', $token[sizeof($token)-1]);
+
+        if($video->original_filename)
         {
-            $finder = Crypt::decrypt($id);
-
-            Coupon::find($finder)->delete();
-
-            return back()->with([
-                'type' => 'success',
-                'message' => 'Coupon deleted successfully.'
-            ]);
+            cloudinary()->destroy(config('app.name').'/'.$token2[0]);
         }
 
+        $video->delete();
+
+        $lesson->delete();
+
         return back()->with([
-            'type' => 'danger',
-            'message' => "Field doesn't match, Try Again!"
-        ]); 
+            'type' => 'success',
+            'message' => 'Lesson deleted successfully.'
+        ]);
     }
 
     public function save_coupon($id, Request $request)
@@ -388,5 +427,41 @@ class CMSController extends Controller
             'type' => 'danger',
             'message' => "Field doesn't match, Try Again!"
         ]); 
+    }
+
+    public function store(Request $request)
+    {
+        $request->validate(
+            [
+                'name' => 'required|unique:stores|max:255',
+                'description' => 'required',
+                'link' => 'required',
+                'theme' => 'required'
+            ],
+            [
+                'name.unique' => 'Store name has already been taken, please use another one!',
+            ]
+        );
+
+        if ($request->file('logo')) {
+            $image = $request->file('logo')->store(
+                'uploads/storeLogo/' . \Auth::user()->username,
+                'public'
+            );
+        }
+        $store = new Store();
+        $store->name = $request->name;
+        $store->description = $request->description;
+        $store->link = $request->link;
+        $store->logo = $image;
+        $store->theme = $request->theme;
+        $store->color = '#fff';
+        $store->user_id = \Auth::user()->id;
+        $store->save();
+
+        return back()->with([
+            'type' => 'success',
+            'message' => $request->name . ' store created successfully'
+        ]);
     }
 }
